+++ lib/docker/api/image.rb
@@ -7,13 +7,22 @@ module Docker
         BuildParams = [:dockerfile, :t, :extrahosts, :remote, :q, :nocache, :cachefrom, :pull, :rm, :forcerm, :memory, :memswap, :cpushares, :cpusetcpus, :cpuperiod, :cpuquota, :buildargs, :shmsize, :squash, :labels, :networkmode, :platform, :target, :outputs]
         class Image < Docker::API::Base
 
+            #################################################
+            # Items in this area to be removed before 1.0.0 #
+            #################################################
             def base_path
                 "/images"
             end
 
             def inspect *args
                 return super.inspect if args.size == 0
+                warn  "WARNING: #inspect is deprecated and will be removed in the future, please use #details instead."
                 name = args[0]
+                details(name)
+            end
+            #################################################
+
+            def details name
                 @connection.get(build_path([name, "json"]))
             end
 
@@ -78,7 +87,7 @@ module Docker
             def commit params = {}, body = {}
                 validate Docker::API::InvalidParameter, [:container, :repo, :tag, :comment, :author, :pause, :changes], params
                 validate Docker::API::InvalidRequestBody, Docker::API::CommitBody, body
-                container = Docker::API::Container.new.inspect(params[:container])
+                container = Docker::API::Container.new.details(params[:container])
                 return container if [404, 301].include? container.status
                 body = JSON.parse(container.body)["Config"].merge(body)
                 @connection.request(method: :post, path: build_path("/commit", params), headers: {"Content-Type": "application/json"}, body: body.to_json)