+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Significant change: `#inspect` is now deprecated and was replaced by `#details` in the following classes:
 * `Docker::API::Container`
+* `Docker::API::Image`
 
 The method will be removed in the refactoring phase.